feat [ROW-566] : playground migration from MUI to Blade UI (#212)
* feat: Adoption of blade is done

* feat: Fixed mobile view bug for sidebar

* feat: changed the editor for json display

* upgrade i18nify-js version

* feat: changed string editor, added version to language and additional changes for it.

* merge commit

* feat: Removed usused packages

* feat: Added Locale dropdown

* feat: Updated country list for banks apis

* feat: Update margin of locale dropdown

* feat: Changed mobile menu alignment

* docs: update module descriptions for clarity and consistency

Refactor module descriptions across multiple pages to provide clearer, more consistent, and professional explanations of their functionalities. The updated descriptions emphasize features, benefits, and integration with the i18nify ecosystem, enhancing user understanding and experience.

* docs(phone): update descriptions for phone utilities to be more professional and concise

Refactor the descriptions of phone-related utilities in the i18nify-playground to provide a more professional, consistent, and concise overview of their functionality. This improves clarity and aligns with the i18nify ecosystem's branding and documentation standards.

* docs: update descriptions for date-related utility functions

Improve clarity and consistency in the descriptions of date-related utility functions in the i18nify-playground. The updated descriptions focus on the essential features, use cases, and integration with the i18nify ecosystem, making them more concise and informative for developers.

* docs(geo): update API descriptions for clarity and consistency

Refactor the descriptions of geo-related APIs to provide clearer, more concise, and consistent information. The updated descriptions emphasize the utility, features, and integration with the i18nify ecosystem, improving readability and user understanding.

* docs(getListOfBanks): update description with detailed utility information

Enhance the description of the `getListOfBanks` component to provide a comprehensive overview of its functionality and benefits. This update aims to clarify the purpose and features of the utility for better understanding and integration within the i18nify ecosystem.

* feat: add new modules for phone, date, and geo functionalities

Introduce new modules for handling phone numbers, dates, and geographical data. This includes utilities for phone number validation, formatting, masking, and parsing, as well as date-time operations and geographical data management. The modules are designed to integrate seamlessly with the i18nify ecosystem, providing consistent internationalization capabilities across the application.

* Create stale-scissors-fly.md

* feat: Add Suspense loading text and changed alignment of locale dropdown for mobile view

* feat: Resolved comments

---------

Co-authored-by: RgnDunes <divyansh.singh@razorpay.com>